refactor: create temporary hook for data fetch
